+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/recovery_lineage/local_recovery/recovered.rs
@@ -1,0 +1,50 @@
+use crate::orchestrator::tests::{
+	operator::status::{
+		running_lanes,
+		running_lanes::{FakeTracker, StateStore, fs, orchestrator, state},
+	},
+	recovery_terminal_support,
+};
+
+#[test]
+fn runtime_recovery_records_recovered_provenance_for_fresh_active_worktree() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let issue = recovery_terminal_support::sample_active_issue("In Progress");
+	let tracker = FakeTracker::new(vec![issue.clone()]);
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let worktree_path = config.worktree_root().join(&issue.identifier);
+
+	fs::create_dir_all(&worktree_path).expect("active worktree path should exist");
+	state::write_run_activity_marker(&worktree_path, "run-1", 1)
+		.expect("activity marker should write");
+
+	let marker = state::read_run_activity_marker_snapshot(&worktree_path)
+		.expect("activity marker should load")
+		.expect("activity marker should exist");
+	let observed_at_unix =
+		marker.last_activity_unix_epoch().expect("activity marker should have a stable timestamp");
+	let recovered_state = orchestrator::recover_runtime_state_from_tracker_and_worktrees(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+	)
+	.expect("runtime recovery should succeed");
+	let mapping = state_store
+		.worktree_for_issue(&issue.id)
+		.expect("mapping lookup should succeed")
+		.expect("recovered mapping should exist");
+	let lease = state_store
+		.lease_for_issue(&issue.id)
+		.expect("lease lookup should succeed")
+		.expect("fresh active marker should recover the lease");
+
+	assert!(
+		recovered_state.recoverable_issues.is_empty(),
+		"fresh marker should recover as the run lease instead of a retry queue item"
+	);
+	assert_eq!(mapping.provenance().source(), "runtime_recovered");
+	assert_eq!(mapping.provenance().created_at_unix(), Some(observed_at_unix));
+	assert_eq!(mapping.provenance().updated_at_unix(), Some(observed_at_unix));
+	assert_eq!(lease.run_id(), "run-1");
+}
