+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/recovery_lineage/local_recovery/worktrees.rs
@@ -1,0 +1,73 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, FakeTracker, StateStore, fs, orchestrator,
+};
+
+#[test]
+fn operator_status_snapshot_includes_local_recovery_worktree_directories() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let worktree_path = config.worktree_root().join("PUB-199");
+
+	fs::create_dir_all(&worktree_path).expect("worktree directory should exist");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+
+	assert_eq!(snapshot.worktrees.len(), 1);
+	assert_eq!(snapshot.worktrees[0].issue_id, "PUB-199");
+	assert!(!snapshot.worktrees[0].branch_name.is_empty());
+	assert_eq!(snapshot.worktrees[0].worktree_path, ".worktrees/PUB-199");
+	assert_eq!(snapshot.worktrees[0].ownership, "cleanup_only");
+	assert!(snapshot.worktrees[0].ownership_reason.contains("local cleanup only"));
+	assert_eq!(snapshot.projects[0].retained_worktree_count, 1);
+}
+
+#[test]
+fn completed_retained_worktree_without_post_review_owner_is_cleanup_only() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue_with_sort_fields(
+		"issue-cleanup",
+		"PUB-199",
+		"Done",
+		&[],
+		Some(4),
+		"2026-03-13T07:16:17.133Z",
+	);
+	let tracker = FakeTracker::new(vec![issue.clone()]);
+	let worktree_path = config.worktree_root().join(&issue.identifier);
+
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-199",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("snapshot should build");
+	let rendered = orchestrator::render_operator_status(&snapshot);
+	let snapshot_json = serde_json::to_value(&snapshot).expect("snapshot should serialize");
+
+	assert!(snapshot.post_review_lanes.is_empty());
+	assert_eq!(snapshot.projects[0].retained_worktree_count, 1);
+	assert_eq!(snapshot.worktrees[0].issue_identifier.as_deref(), Some("PUB-199"));
+	assert_eq!(snapshot.worktrees[0].issue_state.as_deref(), Some("Done"));
+	assert_eq!(snapshot.worktrees[0].ownership, "cleanup_only");
+	assert!(snapshot.worktrees[0].ownership_reason.contains("Issue is Done"));
+	assert_eq!(snapshot_json["worktrees"][0]["ownership"], "cleanup_only");
+	assert_eq!(snapshot_json["worktrees"][0]["issue_state"], "Done");
+	assert!(rendered.contains("role: cleanup_only"));
+	assert!(rendered.contains("reason: Issue is Done"));
+	assert!(!rendered.contains("role: post_review_lane"));
+	assert!(!rendered.contains("classification: blocked"));
+	assert!(!rendered.contains("review_handoff_missing"));
+}
